% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_package_name.R
\name{change_package_name}
\alias{change_package_name}
\title{Change PACKAGE_NAME}
\usage{
change_package_name(x, dev = FALSE, user = NULL)
}
\arguments{
\item{x}{character of vector}

\item{dev}{Development Site vs. not?}

\item{user}{User for the repositories}
}
\value{
Chnage the PROJECT_NAME to the user
}
\description{
Change PACKAGE_NAME
}
