% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapser}
\alias{collapser}
\title{Fixes the multiple suggests}
\usage{
collapser(desc, cn)
}
\arguments{
\item{desc}{dcf file}

\item{cn}{Dependencies tags}
}
\value{
A dcf structure with collapsed dependecies
}
\description{
Fixes the multiple suggests
}
