% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_package.R
\name{example_package}
\alias{example_package}
\title{Make Example Package for testing}
\usage{
example_package(package = c("extrantsr", "cifti", "spotgear"))
}
\arguments{
\item{package}{packages in \code{neuroc.deps} that have an example}
}
\value{
List of a directory and a DESCRIPTION file
}
\description{
Returns a path with a DESCRIPTOIN file in there
}
\examples{
example_package(package = "cifti")
}
