% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ci_fields.R
\name{extract_ci_fields}
\alias{extract_ci_fields}
\title{Extract field from Continuous Integration YAML}
\usage{
extract_ci_fields(file, fields = c("r_binary_packages"))
}
\arguments{
\item{file}{YAML file to read in}

\item{fields}{Fields to extract from YAML}
}
\value{
Either NULL or a list of fields
}
\description{
Extract field from Continuous Integration YAML
}
\examples{
\dontrun{
extract_ci_fields(file = ".travis.yml")
}
}
