% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_desc_remotes.R
\name{fix_desc_remotes}
\alias{fix_desc_remotes}
\title{Fix DESCRIPTION Remotes field}
\usage{
fix_desc_remotes(file, outfile = file)
}
\arguments{
\item{file}{path to DESCRIPTION file}

\item{outfile}{path to DESCRIPTION output file}
}
\value{
Character path of output file
}
\description{
Fix DESCRIPTION Remotes field
}
\examples{
tfile = tempfile()
path = system.file("bad_remotes_DESCRIPTION",
package = "neuroc.deps")
fixed = fix_desc_remotes(path, tfile)
}
