% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repo_install_order.R
\name{get_repo_dep_mat}
\alias{get_repo_dep_mat}
\alias{install_order}
\alias{get_repo_install_order}
\title{Get Install Order of Repositories}
\usage{
get_repo_dep_mat(
  repos,
  dep_type = c("Depends", "Imports", "Suggests"),
  force = FALSE,
  url = "https://api.github.com",
  ...
)

install_order(dep_mat)

get_repo_install_order(repos, dep_type = c("Depends", "Imports", "Suggests"))
}
\arguments{
\item{repos}{Repositories to find the install order}

\item{dep_type}{Dependency types to check against}

\item{force}{should this stop (\code{FALSE}) on missing DESCRIPTION files?}

\item{url}{URL to use to get DESCRIPTION file, passed to
\code{\link[ghtravis]{get_remote_package_dcf}}.}

\item{...}{not used}

\item{dep_mat}{Logial Dependency matrix}
}
\value{
List of dependency order
}
\description{
Get
}
\examples{
\dontrun{
repos = get_repo_names(username = "neuroconductor")
repos = paste0("neuroconductor/", repos)
dep_mat = get_repo_dep_mat(repos)
ord1 = install_order(dep_mat)
order = get_repo_install_order(repos)
}
}
