% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repo_names.R
\name{get_repo_names}
\alias{get_repo_names}
\title{Get Repository names}
\usage{
get_repo_names(...)
}
\arguments{
\item{...}{arguments to be passed to \code{\link{get_all_repos}}}
}
\value{
Vector of repository names
}
\description{
Simple extractor of the \code{names} field from the
GitHub repository
}
