% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_add_collaborator.R
\name{gh_add_collaborator}
\alias{gh_add_collaborator}
\title{Add GitHub Collaborator}
\usage{
gh_add_collaborator(
  repo = "neuroconductor/ANTsR",
  username = "muschellij2",
  .token = NULL,
  ...
)
}
\arguments{
\item{repo}{Repository Name, in the form of user/repo}

\item{username}{Username to add as a collaborator}

\item{.token}{Github personal authentication token (PAT)}

\item{...}{additional arguments to \code{\link{gh}}}
}
\value{
Object of class \code{gh_response}
}
\description{
Add GitHub Collaborator
}
\examples{
\dontrun{
gh_add_collaborator(
repo = "neuroconductor/ANTsR",
username = "muschellij2")
}

}
