% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_ci_fields.R
\name{merge_ci_fields}
\alias{merge_ci_fields}
\title{Merge CI fields}
\usage{
merge_ci_fields(file, template, ...)
}
\arguments{
\item{file}{File to extract fields from}

\item{template}{Character string of template to merge with}

\item{...}{options to \code{\link{extract_ci_fields}}}
}
\value{
Character String of template
}
\description{
Merge CI fields
}
\examples{
    template_file = neuroc_ci_template_path(
    ci = "travis",
    ants = FALSE,
    user = "neuroconductor")
    template = readLines(template_file)
    file = neuroc_example_travis(package = "mimosa")
    file = file$travis_file
    out = merge_ci_fields(file = file, template = template)

}
