% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_ci_template.R
\name{neuroc_ci_template}
\alias{neuroc_ci_template}
\alias{neuroc_travis_template}
\alias{neuroc_appveyor_template}
\title{Neuroconductor CI Template}
\usage{
neuroc_ci_template(path = "DESCRIPTION", ci = c("travis", "appveyor"), ...)

neuroc_travis_template(...)

neuroc_appveyor_template(...)
}
\arguments{
\item{path}{path to DESCRIPTION file}

\item{ci}{Which continuous integration system}

\item{...}{Additional arguments to pass to
\code{\link{neuroc_ci_template_path}}}
}
\value{
File path of YAML file
}
\description{
Neuroconductor CI Template
}
\examples{

}
