% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_require_ants.R
\name{neuroc_dep_ants}
\alias{neuroc_dep_ants}
\title{Neuroconductor ANTsR Dependency}
\usage{
neuroc_dep_ants(release = c("stable", "current"), dev = FALSE,
  bin_packages = c("ITKR", "ANTsR", "ANTsRCore"), table_path = NULL)
}
\arguments{
\item{release}{Stable or development version}

\item{dev}{Development version or not?}

\item{bin_packages}{Packages that require binaries}

\item{table_path}{path to table of packages}
}
\value{
Character vector of packages
}
\description{
Returns the packages in Neuroconductor that require a
set of binary installers
}
\examples{
\dontrun{
neuroc_dep_ants()
}
}
