% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_deps_commit_id.R
\name{neuroc_deps_commit_id}
\alias{neuroc_deps_commit_id}
\title{Get \code{neuroc.deps} commit id}
\usage{
neuroc_deps_commit_id()
}
\value{
A character string
}
\description{
Get \code{neuroc.deps} commit id
}
\examples{
neuroc_deps_commit_id()
}
