% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_desc.R
\name{neuroc_desc}
\alias{neuroc_desc}
\title{Neuroc DESCRIPTION File}
\usage{
neuroc_desc(
  path = "DESCRIPTION",
  table_path = NULL,
  release = c("stable", "current"),
  dev = FALSE,
  verbose = TRUE,
  user = NULL,
  deployment = FALSE,
  force = deployment,
  fix_remotes = TRUE,
  ...
)
}
\arguments{
\item{path}{path to DESCRIPTION file}

\item{table_path}{Path to the table of packages for neuroconductor}

\item{release}{Stable or development version}

\item{dev}{Development Site vs. not?}

\item{verbose}{print diagnostic messages}

\item{user}{GitHub username for repos}

\item{deployment}{indicator if this is a release, not standard running.
Just deployment.}

\item{force}{should this stop (\code{FALSE}) on missing DESCRIPTION files?
Passed to \code{\link{get_repo_dep_mat}}.}

\item{fix_remotes}{Run \code{\link{fix_desc_remotes}} on the DESCRIPTION
file before running.}

\item{...}{arguments to pass to \code{\link{neuroc_install_order}}.}
}
\value{
Path to new DESCRIPTION file
}
\description{
Neuroc DESCRIPTION File
}
\details{
The function will parse a description, delete any remotes for packages in
neuroconductor, add the neuroconductor remotes (in order),
adds \code{biocViews} field (for Travis), puts the \code{covr} in Suggests for
code coverage if not already in Imports/Suggests/Depends,
replaces the SystemRequirements field with the suggested value(s) from
https://goo.gl/x7rcCD
}
