% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_package.R
\name{neuroc_example_description}
\alias{neuroc_example_description}
\title{Make Example DESCRIPTION file for testing}
\usage{
neuroc_example_description(package = c("extrantsr", "cifti", "spotgear",
  "ITKR", "mimosa"))
}
\arguments{
\item{package}{packages in \code{neuroc.deps} that have an example}
}
\value{
List of a directory and a DESCRIPTION file
}
\description{
Returns a path with a DESCRIPTION file in there
}
\examples{
neuroc_example_description(package = "cifti")
}
