% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_package.R
\name{neuroc_example_travis}
\alias{neuroc_example_travis}
\title{Make Example travis file}
\usage{
neuroc_example_travis(package = c("mimosa"))
}
\arguments{
\item{package}{packages in \code{neuroc.deps} that have an example}
}
\value{
List of a directory and a travis file
}
\description{
Returns a path with a travis file in there
}
\examples{
neuroc_example_travis(package = "mimosa")
}
