% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_install_order.R
\name{neuroc_install_order}
\alias{neuroc_install_order}
\title{Neuroconductor Package Install Order}
\usage{
neuroc_install_order(path = "https://neuroconductor.org/neurocPackages",
  release = c("stable", "current"))
}
\arguments{
\item{path}{Path to the table of package}

\item{release}{Stable or development version}
}
\value{
\code{list} of package install order, the dependency matrix,
and repos parsed
}
\description{
Returns the order of Neuroconductor packages installation
}
\examples{
\dontrun{
neuroc_install_order()
}
}
