% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_sha_check.R
\name{neuroc_sha_check}
\alias{neuroc_sha_check}
\title{Check SHA consistency of the system}
\usage{
neuroc_sha_check(release = c("stable", "current"), dev = FALSE,
  table_path = NULL, user = NULL, drop_packages = NULL,
  verbose = FALSE)
}
\arguments{
\item{release}{Stable or development version}

\item{dev}{Development version or not?}

\item{table_path}{path to table of packages}

\item{user}{GitHub username for repos}

\item{drop_packages}{Packages to drop from consistency check}

\item{verbose}{print diagnostic messages}
}
\value{
A list of \code{data.frame}s of information of
dependency commits and the level of dependency
}
\description{
Check SHA consistency of the system
}
\examples{
if (requireNamespace("dplyr", quietly = TRUE)) {
library(dplyr)
L = neuroc_sha_check(release = "stable", dev = FALSE,
drop_packages = "rtapas", verbose = TRUE)
xdf = L$original_df
res = L$dependency_df
res = res \%>\%
  arrange(original_package, level, remote_package)

res = res \%>\%
  group_by(remote_package) \%>\%
  mutate(overall_n_commit = length(unique(remote_commit_id))) \%>\%
  ungroup

res = res \%>\%
  group_by(original_package, remote_package) \%>\%
  mutate(n_commit = length(unique(remote_commit_id)))

res = right_join(xdf \%>\%
                   select(original_package, original_commit_id),
                 res)
bad = res \%>\%
  select(-level) \%>\%
  distinct() \%>\%
  filter(n_commit > 1 | overall_n_commit > 1)
}
}
