% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_table_path.R
\name{neuroc_table_path}
\alias{neuroc_table_path}
\title{Path to Neuroconductor Remote Commits}
\usage{
neuroc_table_path(table_path = NULL, dev = FALSE, user = NULL)
}
\arguments{
\item{table_path}{Path to the table.  If unspecified, uses defaults}

\item{dev}{Whether development should be used or not}

\item{user}{User for the repositories}
}
\value{
Character Vector
}
\description{
Gets the correct table if none are specified (supposed to be robust for
OSLER)
}
\examples{
neuroc_table_path(dev = TRUE)
neuroc_table_path(dev = FALSE)
neuroc_table_path(table_path = "blah")
neuroc_table_path(table_path = "blah", dev = TRUE)
neuroc_table_path(dev = TRUE, user = "oslerinhealth")
neuroc_table_path(dev = FALSE, user = "oslerinhealth")
}
