% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_user.R
\name{neuroc_user}
\alias{neuroc_user}
\title{Neuroconductor GitHub Username}
\usage{
neuroc_user(user = NULL, dev = FALSE)
}
\arguments{
\item{user}{GitHub username}

\item{dev}{Whether development should be used or not}
}
\value{
Character Vector of username
}
\description{
Gets the correct user if none are specified (supposed to be robust for
OSLER)
}
\examples{
neuroc_user(user = "neuroconductor")
neuroc_user(user = "neuroconductor-devel", dev = TRUE)
neuroc_user(dev = TRUE)
neuroc_user(user = "oslerinhealth")
neuroc_user(user = "oslerinhealth", dev = TRUE)
}
