% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_package_name.R
\name{set_env_package_name}
\alias{set_env_package_name}
\title{Set PACKAGE_NAME Environment Variable}
\usage{
set_env_package_name(x, dev = FALSE, user = NULL, deployment = FALSE)
}
\arguments{
\item{x}{character of vector}

\item{dev}{Development Site vs. not?}

\item{user}{User for the repositories}

\item{deployment}{indicator if this is a release, not standard running.
Just deployment.}
}
\value{
Change the PROJECT_NAME to the user
}
\description{
Set PACKAGE_NAME Environment Variable
}
