% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_package_name.R, R/use_neuroc_template.R
\name{change_package_name}
\alias{change_package_name}
\alias{use_neuroc_template}
\alias{use_neuroc_travis_template}
\alias{use_neuroc_appveyor_template}
\alias{neuroc_key}
\alias{add_neuroc_keys}
\alias{remove_neuroc_packages}
\title{Use Neuroconductor Template}
\usage{
change_package_name(x, dev = FALSE, user = NULL)

use_neuroc_template(path = "DESCRIPTION", ci = c("travis", "appveyor"),
  dev = FALSE, table_path = NULL, release = c("stable", "current"),
  bin_packages = c("ITKR", "ANTsR", "ANTsRCore"), verbose = TRUE,
  user = NULL, ...)

use_neuroc_travis_template(...)

use_neuroc_appveyor_template(...)

neuroc_key(ci = c("travis", "appveyor"), dev = FALSE, user = NULL)

add_neuroc_keys(template_file, ci = c("travis", "appveyor"), dev = FALSE,
  user = NULL)

remove_neuroc_packages(..., keep_packs = c("neuroc.deps", "ghtravis",
  "neurocInstall"))
}
\arguments{
\item{dev}{Development Site vs. not?}

\item{user}{User for the repositories}

\item{path}{path to DESCRIPTION file}

\item{ci}{Which continuous integration system}

\item{table_path}{Path to the table of packages for neuroconductor}

\item{release}{Stable or development version}

\item{bin_packages}{Binaries packages required}

\item{verbose}{print diagnostic messages}

\item{...}{Additional arguments to pass to
\code{\link{neuroc_ci_template_path}}}

\item{template_file}{file to template}

\item{keep_packs}{Packages to keep when removing all neuroc packages}
}
\value{
Copy the template to the current directory
}
\description{
Use Neuroconductor Template
}
