% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_ci_template_path.R
\name{neuroc_ci_template_path}
\alias{neuroc_ci_template_path}
\alias{neuroc_travis_template_path}
\alias{neuroc_appveyor_template_path}
\title{Neuroconductor CI Template Paths}
\usage{
neuroc_ci_template_path(ci = c("travis", "appveyor"), ants = FALSE,
  dev = FALSE, user = NULL, deployment = FALSE)

neuroc_travis_template_path(...)

neuroc_appveyor_template_path(...)
}
\arguments{
\item{ci}{Which continuous integration system}

\item{ants}{Does the package depend on ANTs or ANTsR}

\item{dev}{Development Site vs. not?}

\item{user}{GitHub username for repos}

\item{deployment}{indicator if this is a release, not standard running.
Just deployment.}

\item{...}{not used}
}
\value{
File path of YAML file
}
\description{
Neuroconductor CI Template Paths
}
\examples{
neuroc_ci_template_path()
neuroc_appveyor_template_path()
neuroc_ci_template_path(ants = TRUE)
neuroc_travis_template_path()
neuroc_travis_template_path(ants = TRUE)
neuroc_appveyor_template_path()
neuroc_appveyor_template_path(ants = TRUE)
}
