% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_dep_ants.R
\name{neuroc_dep_ants}
\alias{neuroc_dep_ants}
\title{Neuroconductor ANTsR Dependency}
\usage{
neuroc_dep_ants(release = c("stable", "current"), dev = FALSE,
  bin_packages = c("ITKR", "ANTsR", "ANTsRCore"), table_path = NULL,
  deployment = FALSE, user = NULL, force = FALSE, ...)
}
\arguments{
\item{release}{Stable or development version}

\item{dev}{Development version or not?}

\item{bin_packages}{Packages that require binaries}

\item{table_path}{path to table of packages}

\item{deployment}{indicator if this is a release, not standard running.
Just deployment.}

\item{user}{GitHub username for repos}

\item{force}{should this stop (\code{FALSE}) on missing DESCRIPTION files?
Passed to \code{\link{get_repo_dep_mat}}.}

\item{...}{arguments to pass to \code{\link{neuroc_dep_mat}}}
}
\value{
Character vector of packages
}
\description{
Returns the packages in Neuroconductor that require a
set of binary installers
}
\examples{
\dontrun{
neuroc_dep_ants()
}
}
