% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_install_order.R
\name{neuroc_install_order}
\alias{neuroc_install_order}
\alias{neuroc_dep_mat}
\title{Neuroconductor Package Install Order}
\usage{
neuroc_install_order(release = c("stable", "current"), dev = FALSE,
  table_path = NULL, user = NULL)

neuroc_dep_mat(release = c("stable", "current"), dev = FALSE,
  table_path = NULL, user = NULL)
}
\arguments{
\item{release}{Stable or development version}

\item{dev}{Development version or not?}

\item{table_path}{path to table of packages}

\item{user}{GitHub username for repos}
}
\value{
\code{list} of package install order, the dependency matrix,
and repos parsed
}
\description{
Returns the order of Neuroconductor packages installation
}
\examples{
\dontrun{
neuroc_install_order()
}
}
