% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroc_require_ants.R
\name{neuroc_require_ants}
\alias{neuroc_require_ants}
\title{ANTsR Dependency}
\usage{
neuroc_require_ants(
  path = "DESCRIPTION",
  dev = FALSE,
  table_path = NULL,
  release = c("stable", "current"),
  bin_packages = c("ITKR", "ANTsR", "ANTsRCore"),
  verbose = TRUE,
  user = NULL,
  deployment = FALSE,
  force = deployment
)
}
\arguments{
\item{path}{path to DESCRIPTION file}

\item{dev}{Development version or not?}

\item{table_path}{path to table of packages}

\item{release}{Stable or development version}

\item{bin_packages}{Packages that require binaries}

\item{verbose}{print diagnostic messages}

\item{user}{GitHub username for repos}

\item{deployment}{indicator if this is a release, not standard running.
Just deployment.}

\item{force}{should this stop (\code{FALSE}) on missing DESCRIPTION files?
Passed to \code{\link{get_repo_dep_mat}}.}
}
\description{
Returns a logical if binary packages are necessary
}
