% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_neuroc_template.R
\name{use_neuroc_template}
\alias{use_neuroc_template}
\alias{use_neuroc_travis_template}
\alias{use_neuroc_appveyor_template}
\alias{neuroc_key}
\alias{add_neuroc_keys}
\alias{remove_neuroc_packages}
\title{Use Neuroconductor Template}
\usage{
use_neuroc_template(path = "DESCRIPTION", ci = c("travis", "appveyor"),
  dev = FALSE, table_path = NULL, release = c("stable", "current"),
  bin_packages = c("ITKR", "ANTsR", "ANTsRCore"), verbose = TRUE,
  user = NULL, deployment = FALSE, merge_ci = FALSE,
  force = deployment, fix_remotes = TRUE, ...)

use_neuroc_travis_template(...)

use_neuroc_appveyor_template(...)

neuroc_key(ci = c("travis", "appveyor"), dev = FALSE,
  release = FALSE, user = NULL, deployment = FALSE)

add_neuroc_keys(template_file, ci = c("travis", "appveyor"),
  dev = FALSE, user = NULL, deployment = FALSE)

remove_neuroc_packages(..., keep_packs = c("neuroc.deps", "ghtravis",
  "neurocInstall"))
}
\arguments{
\item{path}{path to DESCRIPTION file}

\item{ci}{Which continuous integration system}

\item{dev}{Development Site vs. not?}

\item{table_path}{Path to the table of packages for neuroconductor}

\item{release}{Stable or development version}

\item{bin_packages}{Binaries packages required}

\item{verbose}{print diagnostic messages}

\item{user}{User for the repositories}

\item{deployment}{indicator if this is a release, not standard running.
Just deployment.}

\item{merge_ci}{Should CI fields be merged?  If so,
\code{\link{extract_ci_fields}} is called to grab the relevant fields.
Fields cannot currently be a list}

\item{force}{should this stop (\code{FALSE}) on missing DESCRIPTION files?
Passed to \code{\link{get_repo_dep_mat}}.}

\item{fix_remotes}{Run \code{\link{fix_desc_remotes}} on the DESCRIPTION
file before running.}

\item{...}{Additional arguments to pass to
\code{\link{extract_ci_fields}}}

\item{template_file}{file to template}

\item{keep_packs}{Packages to keep when removing all neuroc packages}
}
\value{
Copy the template to the current directory
}
\description{
Use Neuroconductor Template
}
