% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_neuroc_template.R
\name{use_neuroc_template}
\alias{use_neuroc_template}
\alias{use_neuroc_travis_template}
\alias{use_neuroc_appveyor_template}
\alias{neuroc_key}
\alias{add_neuroc_keys}
\title{Use Neuroconductor Template}
\usage{
use_neuroc_template(path = "DESCRIPTION", ci = c("travis", "appveyor"),
  dev = FALSE, table_path = "https://neuroconductor.org/neurocPackages",
  release = c("stable", "current"), ...)

use_neuroc_travis_template(...)

use_neuroc_appveyor_template(...)

neuroc_key(ci = c("travis", "appveyor"), dev = FALSE)

add_neuroc_keys(template_file, ci = c("travis", "appveyor"), dev = FALSE)
}
\arguments{
\item{path}{path to DESCRIPTION file}

\item{ci}{Which continuous integration system}

\item{dev}{Development Site vs. not?}

\item{table_path}{Path to the table of packages for neuroconductor}

\item{release}{Stable or development version}

\item{...}{Additional arguments to pass to
\code{\link{neuroc_ci_template_path}}}

\item{template_file}{file to template}
}
\value{
Copy the template to the current directory
}
\description{
Use Neuroconductor Template
}
